highlight_evidence.py
import fitz  # PyMuPDF
import sys

def highlight_evidence(pdf_path, output_path, text_to_find):
    """
    Highlights exact matches of text_to_find in the PDF.
    
    Args:
        pdf_path: Path to the source PDF.
        output_path: Path to save the highlighted PDF.
        text_to_find: string to search for and highlight.
    """
    try:
        doc = fitz.open(pdf_path)
    except Exception as e:
        print(f"Error opening PDF: {e}")
        return

    match_count = 0
    
    # Normalize text to find (optional, but good for robust search if needed, 
    # though requirements say 'exact substring')
    # text_to_find = text_to_find.strip()

    print(f"Searching for: '{text_to_find}'...")

    for page_num, page in enumerate(doc):
        # search_for returns a list of Rect objects
        rects = page.search_for(text_to_find)
        
        for rect in rects:
            match_count += 1
            annot = page.add_highlight_annot(rect)
            annot.update() # Ensure annotation is written

    if match_count > 0:
        print(f"Success: Found {match_count} instances")
        try:
            doc.save(output_path)
            print(f"Saved highlighted PDF to: {output_path}")
        except Exception as e:
            print(f"Error saving PDF: {e}")
    else:
        print("ALERT: Text not found in document")
        # Ensure we don't save a file if nothing changed? Or save anyway?
        # Requirement says "Output: Save the modified file to output_path" 
        # but if we didn't modify it, maybe we shouldn't overwrite?
        # I'll save it anyway if user expects an output file regardless of finding highlights 
        # (e.g. pipeline step), but usually if it's an ALERT, maybe no file is better.
        # For now, I'll NOT save to avoid misleading valid outputs, unless requested.
        # Actually, let's save a copy even if no highlights, so the pipeline continues?
        # Re-reading: "If the text is not found: Print ALERT... Output: Save the modified file..."
        # Implies saving even if no highlights (modified=0 edits).
        doc.save(output_path)

if __name__ == "__main__":
    # Test block
    # Create a dummy PDF for testing if it doesn't exist? 
    # Or just use placeholder paths.
    
    # Example usage:
    # python highlight_evidence.py input.pdf output.pdf "some text"
    
    if len(sys.argv) > 3:
        i_path = sys.argv[1]
        o_path = sys.argv[2]
        txt = sys.argv[3]
        highlight_evidence(i_path, o_path, txt)
    else:
        # Default test run as requested
        print("Running test mode with placeholder values (update code or pass args to run real test)...")
        # Since I can't generate a PDF easily without reportlab, I'll just warn.
        print("Usage: python highlight_evidence.py <input_pdf> <output_pdf> <text_to_find>")
