requirements.txt
streamlit
streamlit-pdf-viewer
pypdf
pdfminer.six
google-genai
